Redesign again, use only one buffer for Expr data
This relies on serializing the limbs of large integers into a (unsigned)
char array (the flat tree buffer for all Expr data). Both GMP and Boost
multiprecision (the GMP-independent part) offer functions for this, and
as we expect large rational numbers to be a < 5% use case, we can accept
the detour of copying the limb bytes in and out, for saving the number
in an Expr and re-constructing it for arithmetics.