Store constants in 3 operands for consistency
The `.data.count` on an `Operand` of `Type::constant` should indicate
the number of following operands that belong to it. But we don't want to
restrict constant names to 5 characters (the first in-place `.name`
field). Hence, the first operand carries the header tag and the count,
the second operand is identical to a symbol, and the third identical to
a floating point operand.