Remove sign cache
It's unlikely that querying the sign dynamically is a notable
performance bottleneck. On the contrary, evaluating the sign of a sum is
non-trivial, and doing that in every ctor for a sum might even have a
negative effect, as sign queries are rather sparse.

Also, it simplifies the Expr data structure and tears apart testing of
sign queries.