Re-iterate on Operand layout
We keep the ability to store a zero-terminated string through repeated
`char`s, but enhance the bytes before the main in-place storage (for
floating points, small rationals etc.) with an overlay that can record
the operand count (32bit, we don't need 2**64 - 1 operands anyhow) and
the sign for a serialized large integer.

The `union`s of `Operand` are also named now, which enables an easier
initialization to null with zero instances. Furthermore, the individual
data member `union`s are renamed to `pre`, `mid` and `main` to reflect
their placement.

The new layout allows for constants to be stored in two `Operand`
objects, rather than three.