Re-re-iterator Operand layout
In order to save even more space, we increase the constraints on
names of function and constants: they both must not be longer than 5
non-zero characters. This way, constants are saved within one Operand,
and functions reduce to one Operand plus the arguments.

Note that the naming restrictions on constants and functions are harsh.
They are, however, sufficient for "euler" and "atan2", the longest names
for both types to be present in the default implementation. Also, when
longer names are desired as user extensions, these could be encoded into
smaller ones and decoded by extending the default printing mechanism
(e.g. have a container that maps short names to longer names and prints
the latter).