Store number of child blob instances in Operand
For simple and efficient traversal of logical child operands of
composite expressions, we want to save the number of child `Operand`
instances as part of each composite header.

Due to the size of the `Operand` struct, we limit the number of logical
operands to a 32bit unsigned number - with which we can easily live.