Remove ill-fated polymorphism for children iterations
The idea couldn't work, as something passed as a span of Expr instances
won't be a contiguous sequence of Blobs. Instead, it's a collection of
contiguous sequences of Blobs, where each item in the collection can be
stored anywhere, and the iteration implemented in ChildIterator can
never catch that.

While the idea of having one iteration construct
(ChildIterator/OperandsView) for all cases is nice, it's not possible to
solve this elegantly. Having a ChildIterator with two modes, one for
contiguous Blobs and one for contiguous Expr isn't great and requires a
dispatch on every increment.