Separate Blob/Flag/... details from high-level implementation
This is a patch of considerable size, but it's easier to do this in one
piece. The gist of this is that Blob and its relevant types (Flag, Type)
are only forward declared within the common implementation of the
library. The definition of these details are only visible in
src/lowlevel and tests/lowlevel.

The idea behind this change is to isolate the dangerous parts of the
type design. When the library grows, code in src/ and tests/ will
largely outweigh the lowlevel directories. The nested structure adds
some overhead, but the increase safety should be worth it. In
particular, we can maintain a curated list of accessors and query
functions, as opposed to dispatching over raw Expr header bits
everywhere.